Fixed a null pointer exception issue caused by hookActions being null.
During initialization, this class was reading the MindMapController class to get the hookActions variable from that class but during initialization it was still null.

It is fixed by instead of using a variable to hold the return from the getter
"hookActions = mindMapController.getHookActions();" the code that needs that variable will just call the getter directly so it can have the most up to date variable.